+++ src/dotnet/APIView/APIViewWeb/Languages/PythonLanguageService.cs
@@ -6,7 +6,6 @@ using System.Diagnostics;
 using System.IO;
 using System.Threading.Tasks;
 using ApiView;
-using APIViewWeb.Helpers;
 using Microsoft.ApplicationInsights;
 using Microsoft.Extensions.Configuration;
 
@@ -66,7 +65,7 @@ namespace APIViewWeb
                 _telemetryClient.TrackEvent("Completed Python process run to parse " + originalName);
                 using (var codeFileStream = File.OpenRead(jsonFilePath))
                 {
-                    var codeFile = await CodeFile.DeserializeAsync(codeFileStream, doTreeStyleParserDeserialization: LanguageServiceHelpers.UseTreeStyleParser(this.Name));
+                    var codeFile = await CodeFile.DeserializeAsync(codeFileStream);
                     codeFile.VersionString = VersionString;
                     codeFile.Language = Name;
                     return codeFile;