+++ src/dotnet/APIView/APIViewWeb/Languages/SwaggerLanguageService.cs
@@ -5,7 +5,6 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using ApiView;
-using APIViewWeb.Helpers;
 using Microsoft.ApplicationInsights;
 
 namespace APIViewWeb
@@ -20,13 +19,15 @@ namespace APIViewWeb
 
         public override string ProcessName => throw new NotImplementedException();
 
+        public override bool UsesTreeStyleParser { get; } = false;
+
         public SwaggerLanguageService(TelemetryClient telemetryClient) : base(telemetryClient)
         {
             IsReviewGenByPipeline = true;
         }
         public override async Task<CodeFile> GetCodeFileAsync(string originalName, Stream stream, bool runAnalysis)
         {
-            return await CodeFile.DeserializeAsync(stream, hasSections: true, doTreeStyleParserDeserialization: LanguageServiceHelpers.UseTreeStyleParser(this.Name));
+            return await CodeFile.DeserializeAsync(stream, hasSections: true);
         }
 
         public override string GetProcessorArguments(string originalName, string tempDirectory, string jsonPath)