+++ src/java/apiview-java-processor/src/main/java/com/azure/tools/apiview/processor/model/DiagnosticKind.java
@@ -2,6 +2,7 @@
 package com.azure.tools.apiview.processor.model;
 
 public enum DiagnosticKind {
+    FATAL(4),       // block release!
     ERROR(3),       // red
     WARNING(2),     // yellow
     INFO(1);        // green
@@ -14,6 +15,8 @@ public enum DiagnosticKind {
 
     public static DiagnosticKind fromInt(int level) {
         switch (level) {
+            case 4:
+                return FATAL;
             case 3:
                 return ERROR;
             case 2: