+++ src/java/apiview-java-processor/src/main/java/com/azure/tools/apiview/processor/model/RenderClass.java
@@ -1,0 +1,55 @@
+package com.azure.tools.apiview.processor.model;
+
+import java.util.Arrays;
+import java.util.LinkedHashSet;
+import java.util.Set;
+
+public enum RenderClass {
+    ASSEMBLY("assembly", "java-assembly"),
+
+    CLASS("java-class", "class"),
+    INTERFACE("java-interface", "interface"),
+    ENUM("java-enum", "enum"),
+    ANNOTATION("java-annotation", "annotation"),
+    MODULE_INFO("java-module", "moduleInfo"),
+
+    MODULE_REFERENCE("moduleReference"),
+
+    PUNCTUATION("punctuation"),
+    TYPE_NAME("typeName"),
+    KEYWORD("keyword"),
+    STRING_LITERAL("stringLiteral"),
+    NUMBER("number"),
+    PACKAGE_NAME("java-package", "package", "packageName"),
+    MODULE_NAME("java-module", "moduleName"),
+    ENUM_TYPE("enumType"),
+    ENUM_CONSTANT("enumConstant"),
+    ANNOTATION_NAME("annotationName"),
+    ANNOTATION_PARAMETER_NAME("annotationParameterName"),
+    ANNOTATION_PARAMETER_VALUE("annotationParameterValue"),
+    RETURN_TYPE("returnType"),
+    PARAMETER_TYPE("parameterType"),
+    PARAMETER_NAME("parameterName"),
+    EXTENDS_TYPE("extendsType"),
+    IMPLEMENTS_TYPE("implementsType"),
+    MEMBER_NAME("memberName"),
+    METHOD_NAME("methodName"),
+    FIELD_NAME("fieldName"),
+    MAVEN("java-maven", "maven"),
+    MAVEN_KEY("mavenKey"),
+    MAVEN_VALUE("mavenValue"),
+    MAVEN_DEPENDENCY("dependency"),
+    COMMENT("comment"),
+    JAVADOC("javadoc"),
+    DEPRECATED("deprecated");
+
+    private final Set<String> values;
+
+    RenderClass(String... values) {
+        this.values = new LinkedHashSet<>(Arrays.asList(values));
+    }
+
+    public Set<String> getValues() {
+        return values;
+    }
+}