+++ packages/python-packages/apiview-stub-generator/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## Version 0.3.15 (Unreleased)
+Fixed issue where module-level overloads were not being parsed.
 Update RelatedToLine for empty lines to show in the diff.
 
 ## Version 0.3.14 (2025-02-22)